# specs/001-jwt-auth/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Application API - JWT Authentication
  description: |
    REST API for todo application with JWT-based authentication.
    All protected endpoints require a valid JWT token in the Authorization header.
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Authentication
    description: User authentication operations
  - name: Tasks
    description: Task management operations (protected)

paths:
  /api/v1/auth/session:
    get:
      tags:
        - Authentication
      summary: Get current session
      description: Returns current user session information if authenticated
      operationId: getCurrentSession
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Valid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/users/{user_id}/tasks:
    get:
      tags:
        - Tasks
      summary: List user's tasks
      description: |
        Returns all tasks owned by the specified user.
        Requires JWT token with matching user_id in `sub` claim.
      operationId: listUserTasks
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/userId"
        - name: completed
          in: query
          description: Filter by completion status
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of tasks to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of tasks to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found - user doesn't exist or access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Tasks
      summary: Create a new task
      description: |
        Creates a new task for the specified user.
        Requires JWT token with matching user_id in `sub` claim.
      operationId: createTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found - user doesn't exist or access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

  /api/v1/users/{user_id}/tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: Get a specific task
      description: |
        Returns a specific task by ID.
        Requires JWT token with matching user_id in `sub` claim.
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found - task doesn't exist or access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - Tasks
      summary: Update a task
      description: |
        Updates a specific task. Only provided fields are updated.
        Requires JWT token with matching user_id in `sub` claim.
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found - task doesn't exist or access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: |
        Deletes a specific task.
        Requires JWT token with matching user_id in `sub` claim.
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "204":
          description: Task deleted successfully
        "401":
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found - task doesn't exist or access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by Better Auth.
        Include in Authorization header as: `Bearer <token>`

  parameters:
    userId:
      name: user_id
      in: path
      description: User ID (must match JWT `sub` claim)
      required: true
      schema:
        type: string
        format: uuid
        example: "123e4567-e89b-12d3-a456-426614174000"

    taskId:
      name: task_id
      in: path
      description: Task ID
      required: true
      schema:
        type: string
        format: uuid
        example: "789e0123-e45b-67d8-a901-234567890abc"

  schemas:
    SessionResponse:
      type: object
      required:
        - user_id
        - authenticated
      properties:
        user_id:
          type: string
          format: uuid
          description: Authenticated user ID from JWT
        email:
          type: string
          format: email
          description: User email if available in token
        authenticated:
          type: boolean
          description: Always true for successful response

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title
          example: "Complete project documentation"
        description:
          type: string
          maxLength: 2000
          description: Optional task description
          example: "Write comprehensive API documentation"
        completed:
          type: boolean
          default: false
          description: Initial completion status

    TaskUpdate:
      type: object
      description: Partial update - only include fields to change
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Updated task title
        description:
          type: string
          maxLength: 2000
          description: Updated task description
        completed:
          type: boolean
          description: Updated completion status

    TaskResponse:
      type: object
      required:
        - id
        - owner_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
        owner_id:
          type: string
          format: uuid
          description: User ID of task owner
        title:
          type: string
          description: Task title
        description:
          type: string
          nullable: true
          description: Task description
        completed:
          type: boolean
          description: Whether task is completed
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    TaskListResponse:
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskResponse"
        total:
          type: integer
          description: Total number of tasks (for pagination)
        limit:
          type: integer
          description: Requested limit
        offset:
          type: integer
          description: Requested offset

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Invalid credentials"

    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                description: Location of the error
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Error type

security:
  - bearerAuth: []